+++ src/StateSmith/Runner/Runner.cs
@@ -22,6 +22,7 @@ namespace StateSmith.Runner
     /// </summary>
     public class SmRunner
     {
+        private const string ExpansionVarsPath = "self->vars.";
         RunnerSettings settings;
         public CompilerRunner compilerRunner = new();
         ExceptionPrinter exceptionPrinter;
@@ -89,7 +90,17 @@ namespace StateSmith.Runner
             codeGenContext.mangler = settings.mangler;
             codeGenContext.style = settings.style;
 
-            ConfigReader reader = new ConfigReader(codeGenContext.expander, expansionVarsPath: "self->vars.");
+            foreach (var h in compilerRunner.sm.historyStates)
+            {
+                string actualVarName = codeGenContext.mangler.HistoryVarName(h);
+                codeGenContext.expander.AddVariableExpansion(h.stateTrackingVarName, ExpansionVarsPath + actualVarName);
+                codeGenContext.sm.variables += $"uint8_t {actualVarName};\n";
+                if (h.trackedStateCount > 255) {
+                    throw new VertexValidationException(h, "can't support more than 255 tracked history states right now."); //uint8_t limitation.
+                }
+            }
+
+            ConfigReader reader = new ConfigReader(codeGenContext.expander, expansionVarsPath: ExpansionVarsPath);
             reader.ReadObject(settings.renderConfig);
 
             CBuilder cBuilder = new(codeGenContext);