+++ CHANGELOG.md
@@ -6,8 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## WIP
 ### Added
+- Add shallow history.  https://github.com/StateSmith/StateSmith/issues/63
 - Add `PrefixingModder.cs`
   - Experimental feature to help with duplicate state names. https://github.com/StateSmith/StateSmith/issues/65
+
+### Fixed
+  - Pseudo state transition de-duplication now considers initial state parent incoming transition count too.
+  - Improved exit handling optimization.
 
 ---